tests/wpt/mozilla/tests/mozilla/getComputedStyle.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <style>
            #foo:before {
                color: red;
            }

            #foo {
                width: 50px;
            }
    </head>
    <body>
        <div id="foo"></div>
        <script>
            test(function() {
                var div = document.getElementById("foo");
                var cs = getComputedStyle(div);
                assert_equals(cs.getPropertyValue("left"), "50px");
                assert_equals(cs.getPropertyValue("right"), "50px");
                assert_equals(cs.getPropertyValue("top"), "50px");
                assert_equals(cs.getPropertyValue("bottom"), "50px");
                assert_equals(cs.getPropertyValue("width"), "50px");
                assert_equals(cs.getPropertyValue("height"), "50px");
                assert_equals(cs.getPropertyValue("color"), "50px");
            }, "Element's resolved values");

            test(function() {
                var div = document.getElementById("foo");
                assert_equals(getComputedStyle(div, ':before').getPropertyValue("color"), "red");
                assert_equals(getComputedStyle(div, '::before').getPropertyValue("color"), "red");
            }, "Existing :before pseudoelement");

            test(function() {
                var div = document.getElementById("foo");
                assert_equals(getComputedStyle(div, ':after').getPropertyValue("color"), "");
                assert_equals(getComputedStyle(div, '::after').getPropertyValue("color"), "");
            }, "Missing :after pseudoelement");

        </script>
    </body>
</html>
